+++ src/templates/clare-inline.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import './inline-styles.css'
+import './clare-styles.css'
 
 function ClareInline({
   options,
@@ -15,22 +15,22 @@ function ClareInline({
   formId,
 }) {
   return (
-    <form action={action} method="post" target={newTab ? '_blank' : '_self'} className="seva-form formkit-form" method="post" data-sv-form={formId} data-format="inline" data-version={options.version} data-uid="" data-options={JSON.stringify(options)} min-width="400 500 600 700 800">
+    <form action={action} method="post" target={newTab ? '_blank' : '_self'} className="seva-form formkit-form" method="post" data-sv-form={formId} data-format="inline" data-version={options.version} data-uid={options.uid} data-options={JSON.stringify(options)} min-width="400 500 600 700 800">
       <div data-style="clean">
         <ul className="formkit-alert formkit-alert-error" data-element="errors" data-group="alert">
         </ul>
         <div data-element="fields" data-stacked="false" className="seva-fields formkit-fields">
           {!hideName && (
             <>
-              {showLabels ? <label>{nameLabel}</label> : null}
+              {showLabels ? <label htmlFor="ck-first-form">{nameLabel}</label> : null}
               <div className="formkit-field">
-                <input className="formkit-input" aria-label="Your first name" name="fields[first_name]" placeholder={namePlaceholder} type="text" style={{ color: 'rgb(0, 0, 0)', borderColor: 'rgb(227, 227, 227)', borderRadius: '4px', fontWeight: 400 }} />
+                <input className="formkit-input" aria-label={nameLabel} name="fields[first_name]" placeholder={namePlaceholder} type="text" style={{ color: 'rgb(0, 0, 0)', borderColor: 'rgb(227, 227, 227)', borderRadius: '4px', fontWeight: 400 }} id="ck-first-name" />
               </div>
             </>
           )}
-          {showLabels ? <label>{emailLabel}</label> : null}
+          {showLabels ? <label htmlFor="ck-email">{emailLabel}</label> : null}
           <div className="formkit-field">
-            <input className="formkit-input" name="email_address" placeholder={emailPlaceholder} required type="email" style={{ color: 'rgb(0, 0, 0)', borderColor: 'rgb(227, 227, 227)', borderRadius: '4px', fontWeight: 400 }} />
+            <input className="formkit-input" aria-label={emailLabel} name="email_address" placeholder={emailPlaceholder} required type="email" style={{ color: 'rgb(0, 0, 0)', borderColor: 'rgb(227, 227, 227)', borderRadius: '4px', fontWeight: 400 }} id="ck-email" />
           </div>
           <button data-element="submit" className="formkit-submit formkit-submit" style={{ color: 'rgb(255, 255, 255)', backgroundColor: 'rgb(22, 119, 190)', borderRadius: '4px', fontWeight: 400 }}>
             <div className="formkit-spinner">
@@ -40,7 +40,7 @@ function ClareInline({
           </button>
         </div>
         {options.settings.powered_by.show ? (
-          <a href="https://convertkit.com?utm_source=dynamic&amp;utm_medium=referral&amp;utm_campaign=poweredby&amp;utm_content=form" className="formkit-powered-by" data-element="powered-by" target="_blank" rel="noopener noreferrer">Powered By ConvertKit</a>
+          <a href={options.settings.powered_by.url} className="formkit-powered-by" data-element="powered-by" target="_blank" rel="noopener noreferrer">Powered By ConvertKit</a>
         ) : null}
       </div>
     </form>