+++ src/templates/rainer-inline.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import './rainer-styles.css'
+
+function RainerInline() {
+  return (
+    <form>
+    </form>
+  )
+}
+
+export default RainerInline